requirements.txt
django
djangorestframework
django-cors-headers
drf-spectacular
psycopg2
discord.py
gunicorn
requests
python-dotenv